修改说明

+++ src/misc/xBtnAction.c
@@ -48,7 +48,7 @@ void xBtnConfig(uint8_t debounceMs,uint16_t longPressMs,uint16_t longShootMs,uin
     btnClickTimeOut=clickTimeOutMs/btnDebounceMs;
 }
 
-void xBtnInit(uint16_t id,bool (*getBtnStateFunc)(size_t))
+void xBtnInit(uint16_t id,bool (*getBtnStateFunc)(uint16_t))
 {
     xBtnInfo_t * link = (xBtnInfo_t *)BTN_MALLOC(sizeof(xBtnInfo_t));
     if(link!=NULL)
